fix:eslint and variable names
