+++ src/lib/constants/judge0-languages.ts
@@ -1,30 +1,30 @@
 export interface Judge0Language {
-    slug: string;
+    languageId: string;
     name: string;
     id: number;
 }
 
 export const JUDGE0_LANGUAGES: Judge0Language[] = [
-    { slug: 'python', name: 'Python 3', id: 71 },
-    { slug: 'javascript', name: 'JavaScript', id: 63 },
-    { slug: 'java', name: 'Java', id: 62 },
-    { slug: 'cpp', name: 'C++', id: 54 },
-    { slug: 'c', name: 'C', id: 50 },
-    { slug: 'go', name: 'Go', id: 60 },
-    { slug: 'ruby', name: 'Ruby', id: 72 },
-    { slug: 'typescript', name: 'TypeScript', id: 74 },
-    { slug: 'csharp', name: 'C#', id: 51 },
-    { slug: 'rust', name: 'Rust', id: 73 },
-    { slug: 'swift', name: 'Swift', id: 83 },
-    { slug: 'kotlin', name: 'Kotlin', id: 78 },
+    { languageId: 'python', name: 'Python 3', id: 71 },
+    { languageId: 'javascript', name: 'JavaScript', id: 63 },
+    { languageId: 'java', name: 'Java', id: 62 },
+    { languageId: 'cpp', name: 'C++', id: 54 },
+    { languageId: 'c', name: 'C', id: 50 },
+    { languageId: 'go', name: 'Go', id: 60 },
+    { languageId: 'ruby', name: 'Ruby', id: 72 },
+    { languageId: 'typescript', name: 'TypeScript', id: 74 },
+    { languageId: 'csharp', name: 'C#', id: 51 },
+    { languageId: 'rust', name: 'Rust', id: 73 },
+    { languageId: 'swift', name: 'Swift', id: 83 },
+    { languageId: 'kotlin', name: 'Kotlin', id: 78 },
 ];
 
-export const AVAILABLE_LANGUAGES = JUDGE0_LANGUAGES.map((l) => l.slug);
+export const AVAILABLE_LANGUAGES = JUDGE0_LANGUAGES.map((l) => l.languageId);
 
 export const JUDGE0_LANGUAGE_ID_MAP: Record<string, number> = Object.fromEntries(
-    JUDGE0_LANGUAGES.map((l) => [l.slug, l.id])
+    JUDGE0_LANGUAGES.map((l) => [l.languageId, l.id])
 );
 
 export const JUDGE0_LANGUAGE_NAME_MAP: Record<string, string> = Object.fromEntries(
-    JUDGE0_LANGUAGES.map((l) => [l.slug, l.name])
+    JUDGE0_LANGUAGES.map((l) => [l.languageId, l.name])
 );
