+++ src/lib/hooks/useTaskEditor.ts
@@ -1,5 +1,4 @@
 import { useState, useCallback } from 'react';
-import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 import type {
     BlockNoteContent,
@@ -12,24 +11,22 @@ import { AVAILABLE_LANGUAGES } from '@/lib/constants/judge0-languages';
 import { useMonacoEditor } from '@/lib/hooks/useMonacoEditor';
 
 function convertToEditableTestCases(template: TaskTemplateDetail): EditableTestCase[] {
-    const publicCases = (template.publicTestCases || []).map((tc, idx) => ({
-        id: `public-${idx}`,
-        input: tc.input,
-        output: tc.output,
+    const publicCases = (template.publicTestCases || []).map((testCase, index) => ({
+        id: `public-${index}`,
+        input: testCase.input,
+        output: testCase.output,
         isPublic: true,
     }));
-    const privateCases = (template.privateTestCases || []).map((tc, idx) => ({
-        id: `private-${idx}`,
-        input: tc.input,
-        output: tc.output,
+    const privateCases = (template.privateTestCases || []).map((testCase, index) => ({
+        id: `private-${index}`,
+        input: testCase.input,
+        output: testCase.output,
         isPublic: false,
     }));
     return [...publicCases, ...privateCases];
 }
 
 export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
-    const router = useRouter();
-
     const [title, setTitle] = useState(initialTemplate.title);
     const [description, setDescription] = useState<BlockNoteContent>(
         initialTemplate.description ?? []
@@ -37,7 +34,9 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
     const [tags, setTags] = useState<Tag[]>(initialTemplate.tags ?? []);
     const [selectedLanguages, setSelectedLanguages] = useState<string[]>(() => {
         const languagesFromStarterCodes = [
-            ...new Set(initialTemplate.starterCodes?.map((sc) => sc.language) ?? []),
+            ...new Set(
+                initialTemplate.starterCodes?.map((starterCode) => starterCode.language) ?? []
+            ),
         ];
         return languagesFromStarterCodes.length > 0
             ? languagesFromStarterCodes
@@ -49,13 +48,10 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
     const [starterCodes, setStarterCodes] = useState<StarterCode[]>(
         initialTemplate.starterCodes ?? []
     );
-    const [selectedTestCaseId, setSelectedTestCaseId] = useState<string | null>(
-        testCases.length > 0 ? testCases[0].id : null
-    );
     const [recommendedTimeMinutes, setRecommendedTimeMinutes] = useState(
         initialTemplate.recommendedTimeMinutes ?? 30
     );
-    const [isSaving, setIsSaving] = useState(false);
+    const [isSaving, _setIsSaving] = useState(false); // TODO: Will be used when save functionality is implemented
     const [isRunning, setIsRunning] = useState(false);
     const [testOutput, setTestOutput] = useState('');
 
@@ -65,7 +61,7 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
         currentLanguage,
         handleEditorMount,
         handleLanguageChange: monacoLanguageChange,
-        getCurrentCode,
+        // getCurrentCode, // TODO: Will be used when handleRunCode is implemented
     } = useMonacoEditor({
         initialLanguage: selectedLanguages[0] ?? 'python',
         initialStarterCodes: initialTemplate.starterCodes ?? [],
@@ -79,29 +75,28 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
     const handleAddTestCase = useCallback(
         (seed?: Pick<EditableTestCase, 'input' | 'output' | 'isPublic'>) => {
             const visibilityPrefix = seed?.isPublic === false ? 'private' : 'public';
-            const newId = `${visibilityPrefix}-${Date.now()}`;
+            const newTestCaseId = `${visibilityPrefix}-${Date.now()}`;
             setTestCases((prev) => [
                 ...prev,
                 {
-                    id: newId,
+                    id: newTestCaseId,
                     input: seed?.input ?? '',
                     output: seed?.output ?? '',
                     isPublic: seed?.isPublic ?? true,
                 },
             ]);
-            setSelectedTestCaseId(newId);
-            return newId;
+            return newTestCaseId;
         },
         []
     );
 
     const handleDeleteTestCase = useCallback((id: string) => {
         setTestCases((prev) => {
-            const tc = prev.find((t) => t.id === id);
-            if (!tc) return prev;
+            const testCase = prev.find((testCase) => testCase.id === id);
+            if (!testCase) return prev;
 
-            const publicCount = prev.filter((t) => t.isPublic).length;
-            if (tc.isPublic && publicCount === 1) {
+            const publicCount = prev.filter((testCase) => testCase.isPublic).length;
+            if (testCase.isPublic && publicCount === 1) {
                 toast.error('Cannot delete the only public test case');
                 return prev;
             }
@@ -110,18 +105,17 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
                 return prev;
             }
 
-            const updated = prev.filter((t) => t.id !== id);
-            setSelectedTestCaseId((currentId) =>
-                currentId === id && updated.length > 0 ? updated[0].id : currentId
-            );
-            return updated;
+            const remainingTestCases = prev.filter((testCase) => testCase.id !== id);
+            return remainingTestCases;
         });
     }, []);
 
     const handleUpdateTestCase = useCallback(
         (id: string, field: 'input' | 'output', value: string) => {
             setTestCases((prev) =>
-                prev.map((tc) => (tc.id === id ? { ...tc, [field]: value } : tc))
+                prev.map((testCase) =>
+                    testCase.id === id ? { ...testCase, [field]: value } : testCase
+                )
             );
         },
         []
@@ -129,16 +123,18 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
 
     const handleToggleTestCaseVisibility = useCallback((id: string) => {
         setTestCases((prev) => {
-            const tc = prev.find((t) => t.id === id);
-            if (!tc) return prev;
+            const testCase = prev.find((testCase) => testCase.id === id);
+            if (!testCase) return prev;
 
-            const publicCount = prev.filter((t) => t.isPublic).length;
-            if (tc.isPublic && publicCount === 1) {
+            const publicCount = prev.filter((testCase) => testCase.isPublic).length;
+            if (testCase.isPublic && publicCount === 1) {
                 toast.error('Cannot make the only public test case private');
                 return prev;
             }
 
-            return prev.map((t) => (t.id === id ? { ...t, isPublic: !t.isPublic } : t));
+            return prev.map((testCase) =>
+                testCase.id === id ? { ...testCase, isPublic: !testCase.isPublic } : testCase
+            );
         });
     }, []);
 
@@ -177,7 +173,6 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
         testCases,
         starterCodes,
         currentLanguage,
-        selectedTestCaseId,
         hasUnsavedChanges,
         isSaving,
         testOutput,
@@ -188,7 +183,6 @@ export function useTaskEditor(initialTemplate: TaskTemplateDetail) {
         handleDeleteTestCase,
         handleUpdateTestCase,
         handleToggleTestCaseVisibility,
-        handleSelectTestCase: setSelectedTestCaseId,
         handleEditorMount,
         handleLanguageChange,
         handleRunCode,
